Prevent default break_io being hit
The default break_io value of -1 is being hit when simulated memory at 0xffff is written to.

+++ simulate/z80.cpp
@@ -50,7 +50,7 @@
                           memory->read8((a + 1) & 0xffff)
 
 #define WRITE_RAM(a, b) \
-                          if ((a) == (uint32_t)break_io) \
+                          if ((uint16_t)(a) == (uint32_t)break_io) \
                           { \
                             exit(b); \
                           } \
